Timer & Bug Fixes

+++ src/utility/justANumber.ts
@@ -1,2 +1,2 @@
 //@ts-ignore
-export default (str: string) => str == parseInt(str)
+export default (str: string) => str == parseFloat(str)
